% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRmaf.R
\name{InteRamf}
\alias{InteRamf}
\title{InteRamf}
\usage{
InteRamf()
}
\value{
A webpage to interact with Ramf
}
\description{
InteRamf: a friendly graphical user interface for Ramf package
}
\author{
Marco Chiapello
}
